services/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
temporalio>=1.5.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
supabase>=2.0.0
python-multipart>=0.0.6
google-cloud-storage>=2.14.0
